Add a cargo feature to disable the `getpwuid_r` fallback.
Currently dirs_next and directories_next fall back to calling
`getpwuid_r` when the `HOME` environment variable is not set, on most
Unix-like platforms.

This is undesirable for at least some users, because such behavior is not
specified in the [XDG Base Directory Specification] and differs from the
behavior of the widely used libxdg-basedir. And, calling `getpwuid_r` means
depending on a part of libc which, on some platforms, calls `dlopen` and
runs third-party code that uses filesystem and networking in-process,
which can create problems for some sandboxing scenarios.

This PR puts `getpwuid_r` under control of a feature flag, enabled by
default to maintain compatibility with existing users of the crate, but
which can be disabled by users who don't desire the `getpwuid_r` behavior.

[XDG Base Directory Specification]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html